Transform JS code with ES6 modules
